chore: move submission routes from collectorProfileRoutes to consignRoutes
